+++ fpdf/fpdf.py
@@ -7,6 +7,7 @@
 # * Ported to Python 2.4 by Max (maxpat78@yahoo.it) on 2006-05               *
 # * Maintainer:  Mariano Reingart (reingart@gmail.com) et al since 2008 est. *
 # * Maintainer:  David Alexander (daveankin@gmail.com) et al since 2017 est. *
+# * Maintainer:  Lucas Cimon et al since 2021 est.                           *
 # ****************************************************************************
 """fpdf module (in fpdf package housing FPDF class)
 
@@ -145,6 +146,65 @@ class Annotation(NamedTuple):
     link: Union[str, int] = None
     alt_text: Optional[str] = None
     action: Optional[Action] = None
+    color: Optional[int] = None
+    modification_time: Optional[datetime] = None
+    title: Optional[str] = None
+    quad_points: Optional[tuple] = None
+    page: Optional[int] = None
+
+    def serialize(self, fpdf):
+        "Convert this object dictionnary to a string"
+        rect = (
+            f"{self.x:.2f} {self.y:.2f} "
+            f"{self.x + self.width:.2f} {self.y - self.height:.2f}"
+        )
+
+        out = (
+            f"<</Type /Annot /Subtype /{self.type}"
+            f" /Rect [{rect}] /Border [0 0 0]"
+            # Flag "Print" (bit position 3) specifies to print
+            # the annotation when the page is printed.
+            # cf. https://docs.verapdf.org/validation/pdfa-part1/#rule-653-2
+            f" /F 4"
+        )
+
+        if self.contents:
+            out += f" /Contents {enclose_in_parens(self.contents)}"
+
+        if self.action:
+            out += f" /A <<{self.action.dict_as_string()}>>"
+
+        if self.link:
+            if isinstance(self.link, str):
+                out += f" /A <</S /URI /URI {enclose_in_parens(self.link)}>>"
+            else:  # Dest type ending of annotation entry
+                assert (
+                    self.link in fpdf.links
+                ), f"Link with an invalid index: {self.link} (doc #links={len(fpdf.links)})"
+                out += f" /Dest {fpdf.links[self.link].as_str(fpdf)}"
+
+        if self.color:
+            # pylint: disable=unsubscriptable-object
+            out += f" /C [{self.color[0]} {self.color[1]} {self.color[2]}]"
+
+        if self.title:
+            out += f" /T ({escape_parens(self.title)})"
+
+        if self.modification_time:
+            out += f" /M (D:{self.modification_time:%Y%m%d%H%M%S})"
+
+        if self.quad_points:
+            # pylint: disable=not-an-iterable
+            quad_points = " ".join(
+                f"{quad_point:.2f}" for quad_point in self.quad_points
+            )
+            out += f" /QuadPoints [{quad_points}]"
+
+        if self.page:
+            # Same logic as FPDF._current_page_object_id:
+            out += f" /P {pdf_ref(2 * self.page + 1)}"
+
+        return out + ">>"
 
 
 class TitleStyle(NamedTuple):
@@ -387,6 +447,11 @@ class FPDF(GraphicsStateMixin):
 
         self._drawing_graphics_state_registry = drawing.GraphicsStateDictRegistry()
         self._graphics_state_obj_refs = OrderedDict()
+
+        self.record_text_quad_points = False
+        self.text_quad_points = defaultdict(
+            list
+        )  # page number -> array of 8 × n numbers
 
     @property
     def unifontsubset(self):
@@ -1779,7 +1844,7 @@ class FPDF(GraphicsStateMixin):
         )
 
     @check_page
-    def text_annotation(self, x, y, text):
+    def text_annotation(self, x, y, text, w=1, h=1):
         """
         Puts a text annotation on a rectangular area of the page.
 
@@ -1795,8 +1860,8 @@ class FPDF(GraphicsStateMixin):
                 "Text",
                 x * self.k,
                 self.h_pt - y * self.k,
-                self.k,
-                self.k,
+                w * self.k,
+                h * self.k,
                 contents=text,
             )
         )
@@ -1821,6 +1886,89 @@ class FPDF(GraphicsStateMixin):
                 w * self.k,
                 h * self.k,
                 action=action,
+            )
+        )
+
+    @contextmanager
+    def add_highlight(self, text, title="", color=(1, 1, 0), modification_time=None):
+        """
+        Context manager that adds a single highlight annotation based on the text lines inserted
+        inside its indented block.
+
+        Args:
+            text (str): text of the annotation
+            title (str): the text label that shall be displayed in the title bar of the annotation’s
+                pop-up window when open and active. This entry shall identify the user who added the annotation.
+            color (tuple): a tuple of numbers in the range 0.0 to 1.0, representing a colour used for
+                the title bar of the annotation’s pop-up window
+            modification_time (datetime): date and time when the annotation was most recently modified
+        """
+        if self.record_text_quad_points:
+            raise FPDFException("add_highlight() cannot be nested")
+        self.record_text_quad_points = True
+        yield
+        for page, quad_points in self.text_quad_points.items():
+            self.add_text_markup_annotation(
+                "Highlight",
+                text,
+                quad_points=quad_points,
+                title=title,
+                color=color,
+                modification_time=modification_time,
+                page=page,
+            )
+            self.text_quad_points = defaultdict(list)
+        self.record_text_quad_points = False
+
+    @check_page
+    def add_text_markup_annotation(
+        self,
+        type,
+        text,
+        quad_points,
+        title="",
+        color=(1, 1, 0),
+        modification_time=None,
+        page=None,
+    ):
+        """
+        Adds a text markup annotation on a rectangular area of the page.
+
+        Args:
+            text (str): text of the annotation
+            quad_points (tuple): array of 8 × n numbers specifying the coordinates of n quadrilaterals
+                in default user space that comprise the region in which the link should be activated.
+                The coordinates for each quadrilateral are given in the order: x1 y1 x2 y2 x3 y3 x4 y4
+                specifying the four vertices of the quadrilateral in counterclockwise order
+            title (str): the text label that shall be displayed in the title bar of the annotation’s
+                pop-up window when open and active. This entry shall identify the user who added the annotation.
+            color (tuple): a tuple of numbers in the range 0.0 to 1.0, representing a colour used for
+                the title bar of the annotation’s pop-up window
+            modification_time (datetime): date and time when the annotation was most recently modified
+        """
+        if type not in ("Highlight", "Underline", "Squiggly", "StrikeOut"):
+            raise ValueError(f"Invalid text markup annotation subtype: {type}")
+        if modification_time is None:
+            modification_time = datetime.now()
+        if page is None:
+            page = self.page
+        x_min = min(quad_points[0::2])
+        y_min = min(quad_points[1::2])
+        x_max = max(quad_points[0::2])
+        y_max = max(quad_points[1::2])
+        self.annots[page].append(
+            Annotation(
+                type,
+                contents=text,
+                x=y_min,
+                y=y_max,
+                width=x_max - x_min,
+                height=y_max - y_min,
+                color=color,
+                modification_time=modification_time,
+                title=title,
+                quad_points=quad_points,
+                page=page,
             )
         )
 
@@ -1855,6 +2003,26 @@ class FPDF(GraphicsStateMixin):
         if self.fill_color != self.text_color:
             s = f"q {self.text_color} {s} Q"
         self._out(s)
+        if self.record_text_quad_points:
+            unscaled_width = self.get_normalized_string_width_with_style(
+                txt, self.font_style
+            )
+            if self.font_stretching != 100:
+                unscaled_width *= self.font_stretching / 100
+            w = unscaled_width * self.font_size / 1000
+            h = self.font_size
+            self.text_quad_points[self.page].extend(
+                [
+                    x * self.k,
+                    (self.h - y) * self.k,
+                    (x + w) * self.k,
+                    (self.h - y) * self.k,
+                    x * self.k,
+                    (self.h - y + h) * self.k,
+                    (x + w) * self.k,
+                    (self.h - y + h) * self.k,
+                ]
+            )
 
     @check_page
     def rotate(self, angle, x=None, y=None):
@@ -2181,6 +2349,22 @@ class FPDF(GraphicsStateMixin):
                     f"{x * k:.2f} {(self.h - (y + h)) * k:.2f} m "
                     f"{(x + w) * k:.2f} {(self.h - (y + h)) * k:.2f} l S "
                 )
+
+        if self.record_text_quad_points:
+            x = self.x
+            y = self.y
+            self.text_quad_points[self.page].extend(
+                [
+                    x * self.k,
+                    (self.h - y) * self.k,
+                    (x + w) * self.k,
+                    (self.h - y) * self.k,
+                    x * self.k,
+                    (self.h - y - h) * self.k,
+                    (x + w) * self.k,
+                    (self.h - y - h) * self.k,
+                ]
+            )
 
         s_start = self.x
         s_width, underlines = 0, []
@@ -3127,25 +3311,7 @@ class FPDF(GraphicsStateMixin):
             if page_annots:  # Annotations, e.g. links:
                 annots = ""
                 for annot in page_annots:
-                    # first four things in 'link' list are coordinates?
-                    rect = (
-                        f"{annot.x:.2f} {annot.y:.2f} "
-                        f"{annot.x + annot.width:.2f} {annot.y - annot.height:.2f}"
-                    )
-
-                    # start the annotation entry
-                    annots += (
-                        f"<</Type /Annot /Subtype /{annot.type}"
-                        f" /Rect [{rect}] /Border [0 0 0]"
-                        # Flag "Print" (bit position 3) specifies to print
-                        # the annotation when the page is printed.
-                        # cf. https://docs.verapdf.org/validation/pdfa-part1/#rule-653-2
-                        f" /F 4"
-                    )
-
-                    if annot.contents:
-                        annots += f" /Contents {enclose_in_parens(annot.contents)}"
-
+                    annots += annot.serialize(self)
                     if annot.alt_text is not None:
                         # Note: the spec indicates that a /StructParent could be added **inside* this /Annot,
                         # but tests with Adobe Acrobat Reader reveal that the page /StructParents inserted below
@@ -3153,24 +3319,6 @@ class FPDF(GraphicsStateMixin):
                         self._add_marked_content(
                             self.n, struct_type="/Link", alt_text=annot.alt_text
                         )
-
-                    if annot.action:
-                        annots += f" /A <<{annot.action.dict_as_string()}>>"
-
-                    if annot.link:
-                        if isinstance(annot.link, str):
-                            annots += (
-                                f" /A <</S /URI /URI {enclose_in_parens(annot.link)}>>"
-                            )
-                        else:  # Dest type ending of annotation entry
-                            assert annot.link in self.links, (
-                                f"Page {n} has a link with an invalid index: "
-                                f"{annot.link} (doc #links={len(self.links)})"
-                            )
-                            dest = self.links[annot.link]
-                            annots += f" /Dest {dest.as_str(self)}"
-                    annots += ">>"
-                # End links list
                 self._out(f"/Annots [{annots}]")
             if self.pdf_version > "1.3":
                 self._out("/Group <</Type /Group /S /Transparency" "/CS /DeviceRGB>>")