+++ test/test_annotations.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from fpdf import FPDF
 from fpdf.actions import GoToAction, GoToRemoteAction, LaunchAction, NamedAction
 from fpdf.syntax import DestinationXYZ
-from test.conftest import assert_pdf_equal
+from test.conftest import assert_pdf_equal, EPOCH, LOREM_IPSUM
 
 HERE = Path(__file__).resolve().parent
 
@@ -97,3 +97,26 @@ def test_launch_action(tmp_path):
         h=pdf.font_size,
     )
     assert_pdf_equal(pdf, HERE / "launch_action.pdf", tmp_path)
+
+
+def test_highlighted(tmp_path):
+    pdf = FPDF()
+    pdf.add_page()
+    pdf.set_font("Helvetica", size=24)
+    with pdf.add_highlight("Highlight comment", modification_time=EPOCH):
+        pdf.text(50, 50, "Line 1")
+        pdf.set_y(50)
+        pdf.multi_cell(w=30, txt="Line 2")
+    pdf.cell(w=60, txt="Not highlighted", border=1)
+    assert_pdf_equal(pdf, HERE / "highlighted.pdf", tmp_path)
+
+
+def test_highlighted_over_page_break(tmp_path):
+    pdf = FPDF()
+    pdf.add_page()
+    pdf.set_font("helvetica", size=24)
+    pdf.write(txt=LOREM_IPSUM)
+    pdf.ln()
+    with pdf.add_highlight("Comment", title="Freddy Mercury", modification_time=EPOCH):
+        pdf.write(txt=LOREM_IPSUM)
+    assert_pdf_equal(pdf, HERE / "highlighted_over_page_break.pdf", tmp_path)