+++ test/text/test_multi_cell.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 import fpdf
-from test.conftest import assert_pdf_equal
+from test.conftest import assert_pdf_equal, LOREM_IPSUM
 
 
 HERE = Path(__file__).resolve().parent
@@ -21,39 +21,17 @@ TABLE_DATA = (
 def test_ln_positioning_and_page_breaking_for_multicell(tmp_path):
     doc = fpdf.FPDF(format="letter", unit="pt")
     doc.add_page()
-
     doc.set_font("helvetica", size=TEXT_SIZE)
-    text = (
-        "Lorem ipsum Ut nostrud irure reprehenderit anim nostrud dolore sed "
-        "ut Excepteur dolore ut sunt irure consectetur tempor eu tempor "
-        "nostrud dolore sint exercitation aliquip velit ullamco esse dolore "
-        "mollit ea sed voluptate commodo amet eiusmod incididunt Excepteur "
-        "Excepteur officia est ea dolore sed id in cillum incididunt quis ex "
-        "id aliqua ullamco reprehenderit cupidatat in quis pariatur ex et "
-        "veniam consectetur et minim minim nulla ea in quis Ut in "
-        "consectetur cillum aliquip pariatur qui quis sint reprehenderit "
-        "anim incididunt laborum dolor dolor est dolor fugiat ut officia do "
-        "dolore deserunt nulla voluptate officia mollit elit consequat ad "
-        "aliquip non nulla dolor nisi magna consectetur anim sint officia "
-        "sit tempor anim do laboris ea culpa eu veniam sed cupidatat in anim "
-        "fugiat culpa enim Ut cillum in exercitation magna nostrud aute "
-        "proident laboris est ullamco nulla occaecat nulla proident "
-        "consequat in ut labore non sit id cillum ut ea quis est ut dolore "
-        "nisi aliquip aute pariatur ullamco ut cillum Duis nisi elit sit "
-        "cupidatat do Ut aliqua irure sunt sunt proident sit aliqua in "
-        "dolore Ut in sint sunt exercitation aliquip elit velit dolor nisi "
-    )
-
-    doc.multi_cell(w=144, h=LINE_HEIGHT, border=1, txt=text[:29], ln=0)
-    doc.multi_cell(w=180, h=LINE_HEIGHT, border=1, txt=text[29:60], ln=2)
-    doc.multi_cell(w=144, h=LINE_HEIGHT, border=1, txt=text[60:90], ln=1)
-    doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1, ln=1, txt=text[0:30])
-    doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1, ln=1, txt=text[31:60])
-    doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1, ln=1, txt=text[61:90])
-    doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1, ln=1, txt=text[91:120])
+    doc.multi_cell(w=144, h=LINE_HEIGHT, border=1, txt=LOREM_IPSUM[:29], ln=0)
+    doc.multi_cell(w=180, h=LINE_HEIGHT, border=1, txt=LOREM_IPSUM[29:60], ln=2)
+    doc.multi_cell(w=144, h=LINE_HEIGHT, border=1, txt=LOREM_IPSUM[60:90], ln=1)
+    doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1, ln=1, txt=LOREM_IPSUM[0:30])
+    doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1, ln=1, txt=LOREM_IPSUM[31:60])
+    doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1, ln=1, txt=LOREM_IPSUM[61:90])
+    doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1, ln=1, txt=LOREM_IPSUM[91:120])
     doc.cell(w=72 * 5, h=LINE_HEIGHT, border=1)
     doc.cell(w=1, h=LINE_HEIGHT, ln=2)
-    doc.multi_cell(w=144, h=LINE_HEIGHT, border=1, txt=text[30:90], ln=2)
+    doc.multi_cell(w=144, h=LINE_HEIGHT, border=1, txt=LOREM_IPSUM[30:90], ln=2)
     doc.cell(w=72 * 2, h=LINE_HEIGHT, border=1, ln=2, txt="Lorem ipsum")
     doc.cell(w=72 * 2, h=LINE_HEIGHT, border=1, ln=2, txt="Lorem ipsum")